+++ src/core/injections/controller/checks.py
@@ -171,4 +171,5 @@ def ps_check_failed():
         ps_check = "enter"
       print Back.RED + "(x) Error: '" + ps_check + "' is not a valid answer." + Style.RESET_ALL + "\n"
       pass
+
 #eof